+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## 23.3.4
+## 23.5.0
+BUG FIXES:
+* resource/connector_gcp: fix label key and value do not acccpt only numeric value.
+* resource/volume: disable `enable_deduplication` and `enable_compression` in read function.
+
 NEW FEATURES:
 * resource/cbs: new resource cloud backup service.